+++ src/Daemon/DaemonCommandsHandler.cpp
@@ -144,7 +144,7 @@ bool DaemonCommandsHandler::status(const std::vector<std::string>& args) {
   size_t incoming_connections_count = total_conn - outgoing_connections_count;
   size_t white_peerlist_size = m_srv.getPeerlistManager().get_white_peers_count();
   size_t grey_peerlist_size = m_srv.getPeerlistManager().get_gray_peers_count();
-  uint64_t hashrate = (uint32_t)round(difficulty / CryptoNote::parameters::DIFFICULTY_TARGET);
+  uint32_t hashrate = (uint32_t)round(difficulty / CryptoNote::parameters::DIFFICULTY_TARGET);
   std::time_t uptime = std::time(nullptr) - m_core.getStartTime();
   uint8_t majorVersion = m_core.getBlockMajorVersionForHeight(topBlkIndex);
   bool synced = ((uint32_t)topBlkIndex == (uint32_t)last_known_block_index);
